--- src/main/java/kitchenpos/eatinorders/tobe/EatInOrderType.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorders.tobe;
-
-public enum EatInOrderType {
-    EAT_IN
-}
