+++ src/main/java/kitchenpos/eatinorders/tobe/constant/EatInOrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorders.tobe.constant;
+
+public enum EatInOrderType {
+    EAT_IN
+}
