+++ src/main/java/kitchenpos/eatinorders/tobe/entity/OrderLineItems.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe;
+package kitchenpos.eatinorders.tobe.entity;
 
 import jakarta.persistence.*;
