+++ src/main/java/kitchenpos/eatinorders/tobe/repository/OrderTableRepository.java
@@ -1,0 +1,13 @@
+package kitchenpos.eatinorders.tobe.repository;
+
+import kitchenpos.eatinorders.tobe.entity.OrderTable;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderTableRepository {
+
+    void save(OrderTable orderTable);
+
+    Optional<OrderTable> findBy(UUID id);
+}
