+++ src/main/java/kitchenpos/eatinorders/tobe/service/EatInOrderDomainService.java
@@ -1,0 +1,7 @@
+package kitchenpos.eatinorders.tobe.service;
+
+import java.util.UUID;
+
+public class EatInOrderDomainService {
+
+}
